requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.0
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.27.0
openai==1.17.0
tiktoken==0.6.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
sentence-transformers==2.6.1
faiss-cpu==1.8.0
PyMuPDF==1.24.0
Pillow==10.3.0
python-multipart==0.0.9
pytest==8.1.1
pytest-asyncio==0.23.6
rich==13.7.1
typer==0.11.0
jinja2==3.1.3
aiofiles==23.2.1
